Fix TF provider with weird hack and run `make tfgen`
This is the way pulumi does TF provider wrapping in their own repositories
(adding a git submodule under `upstream`, replacing the path in the `go.mod`
file, and checking out the correct commit in the submodule), so I'm
assuming this is the "proper" way to do it. The reason we need to do this
is because Terraform provider packages follow SemVer but don't follow the way
Go recommends structuring major version bumps in packages, so versions past v1
cannot be pushed to `pkg.go.dev`, where Pulumi downloads the provider from.